funcionamentos de filtragem e ordenacao adicionados, estilizacao tambem
