+++ src/components/CardServico/CardServico.js
@@ -1,27 +1,16 @@
 import React from "react";
-import styled from "styled-components";
-
-
-
-const CardsServico = styled.div`
-    margin-top: 50px;
-    border: solid 1px black;
-    width: 300px;
-    padding: 20px;
-    text-align: left;
-    padding-left: 25px;
-    button {
-        margin: 2px;
-    }
-`
+import { CardsServico } from "./styled";
 
 export default class CardServico extends React.Component {
     render() {
+        const day = new Date(this.props.jobInfos.dueDate).getDay()
+        const month = new Date(this.props.jobInfos.dueDate).getMonth()
+        const year = new Date(this.props.jobInfos.dueDate).getFullYear()
         return(
             <CardsServico>
                     <p>{this.props.jobInfos.title}</p>
                     <p>R${this.props.jobInfos.price.toFixed(2)}</p>
-                    <p>{this.props.jobInfos.dueDate}</p>
+                    <p>{day}/{month}/{year}</p>
                     <button>Ver Detalhes</button>
                     <button>Adicionar no Carrinho</button>
             </CardsServico>
